+++ backend/src/index.ts
@@ -1,13 +1,16 @@
 import cookieParser from "cookie-parser";
+import cors from "cors";
 import express from "express";
 import connectDatabase from "./config/database";
-import { API_VERSION, PORT } from "./config/env";
+import { API_VERSION, CLIENT_URL, PORT } from "./config/env";
 import authRoutes from "./routes/auth.route";
+import userRoutes from "./routes/user.route";
 
 const app = express();
 
 connectDatabase();
 
+app.use(cors({ origin: CLIENT_URL, credentials: true }));
 app.use(cookieParser());
 app.use(express.json());
 
@@ -17,11 +20,13 @@ router.get("/", (_, res) => {
   res.status(200).json({
     success: true,
     message: "API is running",
+    uptime: process.uptime(),
     timestamp: new Date().toISOString(),
   });
 });
 
 router.use("/auth", authRoutes);
+router.use("/user", userRoutes);
 
 app.use("/api/v0", router);