+++ frontend/src/contexts/AuthContext.tsx
@@ -1,0 +1,47 @@
+import apiClient from "@/lib/api";
+import { createContext, useContext, useEffect, useState } from "react";
+
+type AuthContextType = {
+  isAuthenticated: boolean;
+  loading: boolean;
+  checkAuth: () => Promise<void>;
+};
+
+const AuthContext = createContext<AuthContextType>({
+  isAuthenticated: false,
+  loading: false,
+  checkAuth: async () => {},
+});
+
+export function AuthProvider({ children }: { children: React.ReactNode }) {
+  const [isAuthenticated, setIsAuthenticated] = useState(false);
+  const [loading, setLoading] = useState(true);
+
+  const checkAuth = async () => {
+    setLoading(true);
+    try {
+      await apiClient.get("/user/me");
+      setIsAuthenticated(true);
+    } catch {
+      setIsAuthenticated(false);
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  useEffect(() => {
+    checkAuth();
+  }, []);
+
+  return (
+    <AuthContext.Provider value={{ isAuthenticated, loading, checkAuth }}>
+      {children}
+    </AuthContext.Provider>
+  );
+}
+
+// TODO: Find a way to avoid this warning
+// eslint-disable-next-line react-refresh/only-export-components
+export function useAuth() {
+  return useContext(AuthContext);
+}