+++ CHANGELOG.md
@@ -10,8 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Created repository
 - Created backend directory and folder structure
-- Add docker compose yaml with mongodb setup
-- Add auth routes
+- Added docker compose yaml with mongodb setup
+- Added auth routes
+- Created react frontend using vite
+- Added shadcn and react-router to the frontend
+- Added Login and Register routes
+- Added cors middleware to backend
+- Configured axios in frontend
+- Added correct auth communication between frontend and backend
 
 ### Changed